add support for supervanish
this will automatically vanish or unvanish players once they enter spectator mode, regardless on the player permissions. this is useful for server owners wanting staff members to use supervanish without influencing gameplay while on it. this feature is disabled by default and can be enabled by changing "supervanish-hook" in the config.